+++ src/app/api/auth/auth.api.ts
@@ -4,3 +4,11 @@ export const logout = async () => {
   const { data } = await apiInstance.post("/auth/logout");
   return data;
 };
+
+export const deleteAccount = async () => {
+  const refreshToken = localStorage.getItem("refreshToken");
+  const { data } = await apiInstance.delete("/member/me", {
+    data: { refreshToken },
+  });
+  return data;
+};
